Game Loop update
Changed the logic in the part of what scene is going to be next, now its random from 0 to 3. We will be able to load mobs again when the program will detect that we changed scene, resetting the spawner. (which is still not done but i think its a good base.) Though i think variables in the header file is not what i am supposed to do, since i dont see any of you create variables there. So i might need your help to explain to me where should i put it