const char* slander.
it's downright ludicrous to fathom that i, at one point, naively subscribed to the absurdity of such a programming calamity. delimiting one's creative prowess to the constricting grasp of a brain-dead pointer, imprisoned within the stifling boundaries of a predefined scope, stands as a glaring testament to a skill deficiency that is not only conspicuous but also downright pathetic. the revelation that i once championed this abhorrent approach to my code is nothing short of a disheartening plunge into the abyss of regrettable choices and an undeniable testament to the depths of my former programming ignorance.

changes:
- button now uses std::strings instead of it that shall not be named.
- added tracking stuff that can move around. tbh it can be its own class (more abstraction) but its in PlayScene for now.
- almost made a panel
- removed a single blank line in audio.cpp 😌