A memory manager exists...
BUT THEY DON'T WORK lol

everything is allegedly correct besides the global variants of the methods... whenever i uncomment the global versions, it breaks the program.

google says that the overrides NEED to be at the global scope, so i took the scoped versions and made the global override variants
- MemoryManager::operator new
- MemoryManager::operator delete
allegedly exist, but those aren't the global ones that override ALL new and delete's.

possible solutions:
- move the 'scoped versions' to the global scope... but then the variables aren't being tracked which is the whole point of custom overrides for memory allocation/deallocation
- ask Scott for a tip

I've hit the limit of my current memory / 'C' knowledge.

more research is required but idk if this is the time to do that. I still gotta do things that are actually relevant to this project :/

You can check it out tho, its a cool (but scary) class